Update with upcoming TS 4.2 changes
