+++ src/Context.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const Context = createContext<ResizeObserver | null>(null);
