+++ src/index.ts
@@ -1,2 +1,4 @@
-export { ResizeObserverContext, createResizeObserver, Provider } from './ResizeObserverContext';
+export { Context } from './Context';
+export { Provider } from './Provider';
+export { createResizeObserver } from './createResizeObserver';
 export { useResizeObserver } from './useResizeObserver';
